refactor: minimize checkout payload move
