+++ internal/cli/commands_profile_favorites.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"regexp"
 	"sort"
 	"strings"
 
@@ -12,6 +13,8 @@ import (
 	"github.com/mekedron/wolt-cli/internal/service/output"
 	"github.com/spf13/cobra"
 )
+
+var woltVenueIDPattern = regexp.MustCompile(`^[a-fA-F0-9]{24}$`)
 
 type favoritesListOpts struct {
 	lat       float64
@@ -286,7 +289,7 @@ func resolveFavoriteVenueReference(
 	if candidate == "" {
 		return favoriteVenueReference{}, fmt.Errorf("venue id or slug is required")
 	}
-	if domain.LooksLikeObjectID(candidate) {
+	if woltVenueIDPattern.MatchString(candidate) {
 		return favoriteVenueReference{
 			VenueID: strings.ToLower(candidate),
 		}, nil
