+++ internal/cli/objectid.go
@@ -1,4 +1,4 @@
-package domain
+package cli
 
 import (
 	"regexp"
@@ -7,6 +7,6 @@ import (
 
 var objectIDPattern = regexp.MustCompile(`^[a-f0-9]{24}$`)
 
-func LooksLikeObjectID(value string) bool {
+func looksLikeObjectID(value string) bool {
 	return objectIDPattern.MatchString(strings.ToLower(strings.TrimSpace(value)))
 }
